pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[project]
name = "openai-sdk-test"
version = "0.1.0"
description = "CLI for managing vector stores & files"
readme = "README.md"
requires-python = ">=3.12"
tool.uv.package = true
dependencies = [
    "fastapi[standard]>=0.117.1",
    "loguru>=0.7.3",
    "openai>=1.107.3",
    "openai-agents>=0.3.1",
    "pandas>=2.3.2",
    "pip>=25.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-magic-bin>=0.4.14",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "typer>=0.19.1",
]

[project.scripts]
vstore = "openai_vstore_toolkit.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["openai_vstore_toolkit"]
